demo1 server swagger update
