+++ examples/demo1/src/demo-server/go/api_state_transfer.go
@@ -14,5 +14,6 @@ import (
 )
 
 func HandleEvent(w http.ResponseWriter, r *http.Request) {
-	localDBHandleEvent(w, r)
+	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+	w.WriteHeader(http.StatusOK)
 }
