+++ examples/demo1/src/demo-server/go/api_ue_location.go
@@ -13,7 +13,7 @@ import (
 	"net/http"
 )
 
-func PostTrackingNotification(w http.ResponseWriter, r *http.Request) {
-	localDBUpdateTrackedUes(w, r)
-
+func GetUeLocation(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+	w.WriteHeader(http.StatusOK)
 }
