+++ examples/demo1/src/demo-server/go/model_mobility_group_event.go
