+++ examples/demo1/src/demo-server/go/model_subscription_id.go
