+++ examples/demo1/src/demo-server/go/model_user_event_type.go
