+++ examples/demo1/src/demo-server/go/model_zone_info.go
